requirements.txt
foolbox==1.8.0
mlsploit-py==0.1.0
Pillow
torchvision
